feat(plugin): use image cache as plugin
