+++ packages/diagrams-ts/src/providers/create-provider.ts
@@ -8,5 +8,9 @@ export const createProvider = (imagePath: string) => (label: string): Node => ({
     image: imagePath,
     height: 1.4 + 0.2 * ([...label].filter((char) => char === "\n").length + 1),
   },
-  retrieveImage: true,
+  pluginSettings: {
+    localImageCache: {
+      retrieveImage: true,
+    },
+  },
 });
