+++ packages/local-image-cache-plugin/.prettierignore
@@ -1,0 +1,2 @@
+dist
+CHANGELOG.md
