+++ packages/local-image-cache-plugin/image-cache/.gitignore
